Integration test network fixups
Add the creation of a network bridge and associate the vm to the
bridge. This also tries to clean up the libvirtd state between test
runs as currently libvirtd is getting SIGKILL to terminate.

It may be possible to do this without the bridge but it was what I was
able to get working readily while keeping the host side ip assignment
in place.